+++ toolkit/src/commonMain/kotlin/com/airsaid/toolkit/AppInfoExtensions.kt
@@ -1,7 +1,0 @@
-package com.airsaid.toolkit
-
-/**
- * Returns true when [AppInfo.buildType] indicates a debug build.
- */
-val AppInfo.isDebug: Boolean
-  get() = buildType.equals("debug", ignoreCase = true)